Rename package_esp32_index.template.json to package_esp32_index.json
I am trying to use the 'Visual Micro' plugin for Visual Studio and it complain that this file has the 'wrong filename' because of the ".template" bit. Is it possible to change it  so I can use it? And maybe upload a link to this site: https://github.com/arduino/Arduino/wiki/Unofficial-list-of-3rd-party-boards-support-urls so everyone knows you can use Arduino for ESP32 too ?