Create Thread Local Storage for micros()
micros() is using thread specific cycle counts. So, when micros() or delayMicroseconds() are called from different tasks, false rollover events are detected.  The cycle count is task specific, and micros() was using `if(currentCycleCount < lastCycleCount) processRollover()`.

The lastCycleCount variable was updated each time micros() exited with the current Tasks cycle count. So, if the next micros() or delayMicroseconds() was not called from the same Task lastCycleCount had no relation.