use begin() to reset after bus busy
`SCL` and `SDA` need to be detached from the i2c hardware before the i2c hardware is init'd else a glitch is generated.  esp32-hal-i2c.c does not know which pins are assigned to the hardware.  Return the error and let `TwoWire()` do the reinit.

Also, '~' is not the same as '!'  

My Bad.