Increasing stream timeout to avoid read timeout
If using nginx webserver to feed the OTA stream with the binary, i'm getting reproducely _error #6 (stream read timeout) on a ESP32 Pycom LoPy with CPU revision 0.
This error does not occur with some other ESP32 boards with same setup and server, don't know why.

But i could fix the issue with the LoPy following a hint for ESP 8266 Update.cpp, found in this issue:
https://github.com/esp8266/Arduino/issues/1157

This PR is a quick&dirty fix. Problem should be further analysed and then solved at the root cause.