Increase UART concurrency
While data is beginning received, it is stuck in the RX Fifo until the fifo fills to rxfifo_full_thrhd, or there is a break in the continuous data stream longer than the configured rx_tout_thrhd.  This fix changes processing so that if a read() sees an empty queue, or available() is called, the RX fifo is flushed.  This will decrease any latency between reception at the port and data handling in Serial().

This mod has not been exhaustively tested, but I have encountered no issues with my per-existing code.