Remove empty line that trips copyright detection
This file contains an empty line in the middle of its copyright
statement. This is uncommon in this codebase and in general and makes it
difficult to properly detect a copyright. It triggered a rare copyright
detection issue in scancode-toolkit. Remove this one case help users
of this codebase to better scanc and comply with its licenses.

Link: https://github.com/nexB/scancode-toolkit/issues/1565
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>