Handle HTTP Response with no content-length header
I ran into an issue with a Google Rest API which did not send back a content-length header or a transfer-encoding header.  If this happens the HTTPClient gives the user no access to the response body in any way.  So my solution is to set the _size variable to the remaining length of data available on _client which makes sense and works perfectly for me now.