Fix deadlock in BtSerial if connection is closed before Tx ACK (#3657)
For every packet sent, the transmission task waits for ACK. This is
implemented as a blocking semaphore, which is posted by the ESP SPP
callback in the ESP_SPP_WRITE_EVT event, after the data has been
sent. However, if the connection is closed before this event is
notified by the callback, the semaphore is never posted and the task
is locked forever.

This patch fixes the issue by setting a timeout for the transmission
Ack. This timeout may be adjusted at build time by setting the macro
TX_ACK_TO.

Note that in this scenatio, there is no confirmation if the data has
reached its destinatary. We will assume yes, so the tranmission buffer
is cleared, even if the Tx Ack event is not received.