Fix TLS memory leaks after early failures of start_ssl_client()
TLS data structures for CA certificate chain and client certifcate and private key were not deallocated if the function returns earlier due to some error. The fix makes sure to free this data in addition to the other cleanup already present in stop_ssl_socket(), which is called by the client when start_ssl_client() fails.