Added KEY_SPACE definition to USBHIDKeyboard.h
This commit introduces a new key definition for the space key in the USBHIDKeyboard.h file. Previously, there was no direct way to refer to the space key in the HID keyboard interface. Users had to use the ASCII value '0x20', which was not intuitive and inconsistent with how other keys are handled.

By defining `KEY_SPACE` as `0x20`, this commit aligns the handling of the space key with other special keys in the library, making the code more readable and user-friendly. This change simplifies the process of programming keyboard interactions, especially for beginners, by allowing them to use `Keyboard.press(KEY_SPACE);` in a manner consistent with other keys.