feat(debug): pass objdump path to resolve globals
Cortex-debug requires objdump to be in the same folder as gdb, or needs
the full path to the executable in the `launch.json` file. This is now
possible with the new debugging API.

On Arduino-ESP32 core versions 2.x, though, this causes a crash within
the Cortex-Debug extension as a result of loading the symbol table. This
feature is thus disabled for now.